Create RightSection
